+++ src/core.scss
@@ -1,0 +1,4 @@
+@import './styles/vars';
+@import './styles/reset';
+@import './styles/base';
+@import './styles/layout';
