+++ src/footer/footer.scss
@@ -1,4 +1,4 @@
 footer {
-  color: #002a35;
+  // color: #002a35;
   grid-column: 1/4;
 }
