+++ src/form/form.scss
@@ -3,9 +3,6 @@ label {
 }
 .results {
   grid-column: 2/4;
-}
-
-ul {
   border: 2px solid #002a35;
   border-radius: 6px;
   list-style-type: none;
