+++ src/header/Header.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link, NavLink } from 'react-router-dom';
-import './styles/header.scss';
+import './header.scss';
 
 const Header = () => {
   return (
