+++ src/history/History.js
@@ -1,10 +1,12 @@
 import React from 'react';
-import './styles/history.scss'
+import { withRouter } from 'react-router-dom';
+
+import './history.scss'
 
 class History extends React.Component {
 
   render() {
-  let items = this.props.history.map((query, i) => <li key={i}><button onClick={this.props.handler(query)}>{query.method} {query.url}</button></li>);
+    let items = this.props.history.map((query, i) => <li key={i}><button onClick={this.props.handler(query)}>{query.method} {query.url}</button></li>);
 
     return (
       <section id="history">
