+++ src/styles/_base.scss
@@ -1,0 +1,6 @@
+body {
+  font-family: $primary-font;
+  font-size: 1.5vw;
+  background-color: $primary-color;
+  color: $secondary-color
+}
