+++ src/styles/_vars.scss
@@ -1,0 +1,17 @@
+// Global config
+$site-width: 960px;
+$primary-font: 'helvetica';
+
+// Colors
+$primary-color: #e75c4b;
+$secondary-color: #002a35;
+$black: #333;
+$white: #fff;
+
+// Gutters
+$gutter: 2vw;
+$gutter-medium: $gutter / 2;
+$gutter-small: $gutter / 4;
+
+// General styling
+$round-border: 3px;
