--- src/styles/header.scss
@@ -1,10 +1,0 @@
-h1 {
-  background-color: #002a35e1;
-  color: #c3786f;
-  padding: 1rem;
-  margin: 0;
-}
-
-header {
-  grid-column: 1/4;
-}
